Nitrogen:  Silence AE refcon cast warnings in LP64
Casting a 32-bit int to a 64-bit pointer yields a warning.

+++ nitric/Nitrogen/AEObjects.hh
@@ -379,7 +379,7 @@ namespace Nitrogen
 		return AEInstallObjectAccessor( OSLAccessor( desiredClass,
 		                                             containerType,
 		                                             StaticUPP< OSLAccessorUPP, accessor >(),
-		                                             (::SRefCon) accessorRefCon,
+		                                             (::SRefCon) (long) accessorRefCon,
 		                                             isSysHandler ) );
 	}
